requirements.txt
# ================================================================================
# REQUIREMENTS - PROJETO DE ANÁLISE DE DADOS DE SURF
# ================================================================================

# === CORE DEPENDENCIES ===
# dbt with DuckDB adapter
dbt-core>=1.7.0,<2.0.0
dbt-duckdb>=1.8.0,<2.0.0

# Database
duckdb>=0.9.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests for APIs
requests>=2.31.0

# Date and time manipulation
python-dateutil>=2.8.0

# Configuration management
PyYAML>=6.0

# === DATA SCIENCE ===
# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Plotting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# === DEVELOPMENT ===
# Code formatting
black>=23.0.0
isort>=5.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type hints
mypy>=1.5.0

# === OPTIONAL - FOR FUTURE FEATURES ===
# Machine Learning (commented out for now)
# scikit-learn>=1.3.0
# xgboost>=1.7.0

# Streamlit for dashboards (commented out for now)
# streamlit>=1.25.0

# FastAPI for APIs (commented out for now)
# fastapi>=0.100.0
# uvicorn>=0.23.0

# === UTILITIES ===
# Progress bars
tqdm>=4.65.0

# Logging
loguru>=0.7.0

# Environment variables
python-dotenv>=1.0.0
